Initial Commit
This module provides puppet code to create an index that is a
short-term workaround to PDB-3323.  The puppet code will make the
index only if you are on a version of PE 2016.5.

The puppet code will then drop that index if you upgrade to a
version that includes an actual fix to the performance problem
thus insulating the user from an unsupported index causes issues
later.