Namespace postgresql_psql to provide broad compatibility
the onlyif parameter on postgresql_psql is only available on 4.5+.
Since this module needs to be compatible with as many users as
possible having that constraint isn't acceptable.

This commit copies the current postgresql_psql from the master
branch of postgresql into this module and namespaces it so it
won't conflict with user's own version of postgresql.

As a result this module ends up requiring no dependencies.